requirements.txt
flask
pypdf
werkzeug
pandas
Flask-SQLAlchemy
psycopg2-binary
gunicorn
scikit-learn
sentence-transformers
flask-login

